Rename favorite_site.dart -> favorite_sight.dart
