+++ lib/domain/favorite_sight.dart
